pipelines/google_apis/requirements.txt
google_api_python_client==2.70.0
google_auth_oauthlib==0.8.0
gspread==5.7.2
gspread_dataframe==3.3.0
pandas==1.5.2
protobuf==4.21.12
pydrive==1.3.1
